// src/proto/constants.ts
/*
 * SPDX-FileCopyrightText: 2026 Espressif Systems (Shanghai) CO LTD
 *
 * SPDX-License-Identifier: Apache-2.0
 */

/**
 * Generated by the protoc-gen-ts.  DO NOT EDIT!
 * compiler version: 5.28.2
 * source: constants.proto
 * git: https://github.com/thesayyn/protoc-gen-ts */
export enum Status {
  Success = 0,
  InvalidSecScheme = 1,
  InvalidProto = 2,
  TooManySessions = 3,
  InvalidArgument = 4,
  InternalError = 5,
  CryptoError = 6,
  InvalidSession = 7,
}