// src/proto/esp_local_ctrl.ts
export enum LocalCtrlMsgType {
  TypeCmdGetPropertyCount = 0,
  TypeRespGetPropertyCount = 1,
  TypeCmdGetPropertyValues = 4,
  TypeRespGetPropertyValues = 5,
  TypeCmdSetPropertyValues = 6,
  TypeRespSetPropertyValues = 7,
}

/**
 * Command to get property count
 */
export class CmdGetPropertyCount extends pb_1.Message {
  constructor(data?: any[] | {}) {
    super();
    pb_1.Message.initialize(
      this,
      Array.isArray(data) ? data : [],
      0,
      -1,
      [],
      []
    );
  }

  static fromObject(data: {}): CmdGetPropertyCount {
    return new CmdGetPropertyCount(data);
  }

  toObject(): {} {
    return {};
  }

  serialize(): Uint8Array {
    const writer = new pb_1.BinaryWriter();
    return writer.getResultBuffer();
  }

  static deserialize(_bytes: Uint8Array): CmdGetPropertyCount {
    return new CmdGetPropertyCount();
  }

  serializeBinary(): Uint8Array {
    return this.serialize();
  }

  static deserializeBinary(bytes: Uint8Array): CmdGetPropertyCount {
    return CmdGetPropertyCount.deserialize(bytes);
  }
}
// ...
export class CmdGetPropertyValues extends pb_1.Message {
  #one_of_decls: number[][] = [];
  constructor(
    data?:
      | any[]
      | {
          indices?: number[];
        }
  ) {
    super();
    pb_1.Message.initialize(
      this,
      Array.isArray(data) ? data : [],
      0,
      -1,
      [1],
      this.#one_of_decls
    );
    if (!Array.isArray(data) && typeof data == "object") {
      if ("indices" in data && data.indices != undefined) {
        this.indices = data.indices;
      }
    }
  }
  get indices() {
    return pb_1.Message.getFieldWithDefault(this, 1, []) as number[];
  }
  set indices(value: number[]) {
    pb_1.Message.setField(this, 1, value);
  }
  static fromObject(data: { indices?: number[] }): CmdGetPropertyValues {
    const message = new CmdGetPropertyValues({});
    if (data.indices != null) {
      message.indices = data.indices;
    }
    return message;
  }
  toObject() {
    const data: {
      indices?: number[];
    } = {};
    if (this.indices != null) {
      data.indices = this.indices;
    }
    return data;
  }
  serialize(): Uint8Array;
  serialize(w: pb_1.BinaryWriter): void;
  serialize(w?: pb_1.BinaryWriter): Uint8Array | void {
    const writer = w || new pb_1.BinaryWriter();
    if (this.indices.length) writer.writePackedUint32(1, this.indices);
    if (!w) return writer.getResultBuffer();
  }
  static deserialize(bytes: Uint8Array): CmdGetPropertyValues {
    const reader = new pb_1.BinaryReader(bytes);
    const message = new CmdGetPropertyValues();
    while (reader.nextField()) {
      if (reader.isEndGroup()) break;
      switch (reader.getFieldNumber()) {
        case 1:
          message.indices = reader.readPackedUint32();
          break;
        default:
          reader.skipField();
      }
    }
    return message;
  }
  serializeBinary(): Uint8Array {
    return this.serialize();
  }
  static deserializeBinary(bytes: Uint8Array): CmdGetPropertyValues {
    return CmdGetPropertyValues.deserialize(bytes);
  }
}
// ...
export class PropertyValue extends pb_1.Message {
  #one_of_decls: number[][] = [];
  constructor(
    data?:
      | any[]
      | {
          index?: number;
          value?: Uint8Array;
        }
  ) {
    super();
    pb_1.Message.initialize(
      this,
      Array.isArray(data) ? data : [],
      0,
      -1,
      [],
      this.#one_of_decls
    );
    if (!Array.isArray(data) && typeof data == "object") {
      if ("index" in data && data.index != undefined) {
        this.index = data.index;
      }
      if ("value" in data && data.value != undefined) {
        this.value = data.value;
      }
    }
  }
  get index() {
    return pb_1.Message.getFieldWithDefault(this, 1, 0) as number;
  }
  set index(value: number) {
    pb_1.Message.setField(this, 1, value);
  }
  get value() {
    return pb_1.Message.getFieldWithDefault(
      this,
      2,
      new Uint8Array(0)
    ) as Uint8Array;
  }
  set value(value: Uint8Array) {
    pb_1.Message.setField(this, 2, value);
  }
  static fromObject(data: {
    index?: number;
    value?: Uint8Array;
  }): PropertyValue {
    const message = new PropertyValue({});
    if (data.index != null) {
      message.index = data.index;
    }
    if (data.value != null) {
      message.value = data.value;
    }
    return message;
  }
  toObject() {
    const data: {
      index?: number;
      value?: Uint8Array;
    } = {};
    if (this.index != null) {
      data.index = this.index;
    }
    if (this.value != null) {
      data.value = this.value;
    }
    return data;
  }
  serialize(): Uint8Array;
  serialize(w: pb_1.BinaryWriter): void;
  serialize(w?: pb_1.BinaryWriter): Uint8Array | void {
    const writer = w || new pb_1.BinaryWriter();
    if (this.index != 0) writer.writeUint32(1, this.index);
    if (this.value.length) writer.writeBytes(2, this.value);
    if (!w) return writer.getResultBuffer();
  }
  static deserialize(bytes: Uint8Array): PropertyValue {
    const reader = new pb_1.BinaryReader(bytes);
    const message = new PropertyValue();
    while (reader.nextField()) {
      if (reader.isEndGroup()) break;
      switch (reader.getFieldNumber()) {
        case 1:
          message.index = reader.readUint32();
          break;
        case 2:
          message.value = reader.readBytes();
          break;
        default:
          reader.skipField();
      }
    }
    return message;
  }
  serializeBinary(): Uint8Array {
    return this.serialize();
  }
  static deserializeBinary(bytes: Uint8Array): PropertyValue {
    return PropertyValue.deserialize(bytes);
  }
}

export class CmdSetPropertyValues extends pb_1.Message {
  #one_of_decls: number[][] = [];
  constructor(
    data?:
      | any[]
      | {
          props?: PropertyValue[];
        }
  ) {
    super();
    pb_1.Message.initialize(
      this,
      Array.isArray(data) ? data : [],
      0,
      -1,
      [1],
      this.#one_of_decls
    );
    if (!Array.isArray(data) && typeof data == "object") {
      if ("props" in data && data.props != undefined) {
        this.props = data.props;
      }
    }
  }
  get props() {
    return pb_1.Message.getRepeatedWrapperField(
      this,
      PropertyValue,
      1
    ) as PropertyValue[];
  }
  set props(value: PropertyValue[]) {
    pb_1.Message.setRepeatedWrapperField(this, 1, value);
  }
  static fromObject(data: {
    props?: ReturnType<typeof PropertyValue.prototype.toObject>[];
  }): CmdSetPropertyValues {
    const message = new CmdSetPropertyValues({});
    if (data.props != null) {
      message.props = data.props.map((item) => PropertyValue.fromObject(item));
    }
    return message;
  }
  toObject() {
    const data: {
      props?: ReturnType<typeof PropertyValue.prototype.toObject>[];
    } = {};
    if (this.props != null) {
      data.props = this.props.map((item: PropertyValue) => item.toObject());
    }
    return data;
  }
  serialize(): Uint8Array;
  serialize(w: pb_1.BinaryWriter): void;
  serialize(w?: pb_1.BinaryWriter): Uint8Array | void {
    const writer = w || new pb_1.BinaryWriter();
    if (this.props.length)
      writer.writeRepeatedMessage(1, this.props, (item: PropertyValue) =>
        item.serialize(writer)
      );
    if (!w) return writer.getResultBuffer();
  }
  static deserialize(bytes: Uint8Array): CmdSetPropertyValues {
    const reader = new pb_1.BinaryReader(bytes);
    const message = new CmdSetPropertyValues();
    while (reader.nextField()) {
      if (reader.isEndGroup()) break;
      switch (reader.getFieldNumber()) {
        case 1: {
          const itemBytes = reader.readBytes();
          pb_1.Message.addToRepeatedWrapperField(
            message,
            1,
            PropertyValue.deserialize(itemBytes),
            PropertyValue
          );
          break;
        }
        default:
          reader.skipField();
      }
    }
    return message;
  }
  serializeBinary(): Uint8Array {
    return this.serialize();
  }
  static deserializeBinary(bytes: Uint8Array): CmdSetPropertyValues {
    return CmdSetPropertyValues.deserialize(bytes);
  }
}